made PaymentAmount decimal

+++ client/Lykke.Service.PayInternal.Client/Models/PaymentRequest/PaymentRequestOrderModel.cs
@@ -5,7 +5,7 @@ namespace Lykke.Service.PayInternal.Client.Models.PaymentRequest
     public class PaymentRequestOrderModel
     {
         public string Id { get; set; }
-        public double PaymentAmount { get; set; }
+        public decimal PaymentAmount { get; set; }
         public DateTime DueDate { get; set; }
         public DateTime CreatedDate { get; set; }
         public decimal ExchangeRate { get; set; }
